+++ bloco_7/dia_1/.vscode/launch.json
@@ -11,7 +11,7 @@
       "skipFiles": [
         "<node_internals>/**"
       ],
-      "program": "${workspaceFolder}/script.js"
+      "program": "${file}"
     }
   ]
 }
